refactor: apply database connection review changes

+++ src/main.cpp
@@ -8,8 +8,7 @@ std::string getEnvironmentVariable(const char* name) {
     const char* value = std::getenv(name);
 
     if (value == nullptr) {
-        throw std::runtime_error(std::string("Missing environment variable: ") + 
-                                             name);
+        throw std::runtime_error(std::string("Missing environment variable: ") + name);
     }
 
     return std::string(value);
@@ -18,16 +17,16 @@ std::string getEnvironmentVariable(const char* name) {
 int main() {
     try {
         const std::string host = getEnvironmentVariable("DB_HOST");
-        const int port = std::stoi(getEnvironmentVariable("DB_PORT"));
+        const unsigned port = static_cast<unsigned>(std::stoi(getEnvironmentVariable("DB_PORT")));
         const std::string username = getEnvironmentVariable("DB_USER");
         const std::string password = getEnvironmentVariable("DB_PASSWORD");
         const std::string databaseName = getEnvironmentVariable("DB_NAME");
 
         Database database(host,
-            port,
-            username,
-            password,
-            databaseName);
+                          port,
+                          username,
+                          password,
+                          databaseName);
     }
     catch (const std::exception& error) {
         std::cerr << error.what() << std::endl;
